Implements review considerations
